Peyeramid_CodeRework_V1 + SH_V2

+++ Assets/Scripts/Enemy/IEnemy.cs
@@ -1,6 +1,6 @@
 interface IEnemy
 {
     void Attack();
-    void Move(int direction);
+    void Move(float direction);
     void Trigger();
 }
